+++ x-pack/plugin/eql/build.gradle
@@ -20,7 +20,7 @@ dependencies {
   compileOnly(project(':modules:lang-painless:spi'))
   compileOnly project(xpackModule('ql'))
 
-  compileOnly project(xpackModule('qa-util:common'))
+  compileOnly project(path: xpackModule('qa-util:common'))
 
   testImplementation project(':test:framework')
   testImplementation(testArtifact(project(xpackModule('core'))))
