+++ x-pack/plugin/qa-util/build.gradle
@@ -2,4 +2,5 @@ apply plugin: 'elasticsearch.java'
 
 dependencies {
   api project(':test:framework')
+
 }
