+++ x-pack/plugin/sql/src/test/java/org/elasticsearch/xpack/sql/expression/function/scalar/datetime/DateTimeTestUtils.java
@@ -7,7 +7,7 @@
 
 package org.elasticsearch.xpack.sql.expression.function.scalar.datetime;
 
-import org.elasticsearch.xpack.sql.util.DateUtils;
+import org.elasticsearch.xpack.qautil.DateUtils;
 
 import java.time.Clock;
 import java.time.Duration;
@@ -24,11 +24,11 @@ public class DateTimeTestUtils {
     private DateTimeTestUtils() {}
 
     public static ZonedDateTime dateTime(int year, int month, int day, int hour, int minute) {
-        return ZonedDateTime.of(year, month, day, hour, minute, 0, 0, DateUtils.UTC);
+        return ZonedDateTime.of(year, month, day, hour, minute, 0, 0, org.elasticsearch.xpack.qautil.DateUtils.UTC);
     }
 
     public static ZonedDateTime dateTime(int year, int month, int day, int hour, int minute, int seconds, int nanos) {
-        return dateTime(year, month, day, hour, minute, seconds, nanos, DateUtils.UTC);
+        return dateTime(year, month, day, hour, minute, seconds, nanos, org.elasticsearch.xpack.qautil.DateUtils.UTC);
     }
 
     public static ZonedDateTime dateTime(int year, int month, int day, int hour, int minute, int seconds, int nanos, ZoneId zoneId) {
@@ -40,7 +40,7 @@ public class DateTimeTestUtils {
     }
 
     public static OffsetTime time(long millisSinceEpoch) {
-        return DateUtils.asTimeOnly(millisSinceEpoch);
+        return org.elasticsearch.xpack.qautil.DateUtils.asTimeOnly(millisSinceEpoch);
     }
 
     public static OffsetTime time(int hour, int minute, int second, int nano) {
